requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==23.2.0
backoff==1.11.1
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
category-encoders==2.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.2
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.3
django-redis==5.4.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.10.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.2
idna==3.7
ImageHash==4.3.1
imbalanced-learn==0.12.3
imblearn==0.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
liac-arff==2.5.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multimethod==1.12
musicbrainzngs==0.7.1
narwhals==1.13.3
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
phik==0.12.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.0
protobuf==5.28.3
psutil==6.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pylast==5.3.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
RapidFuzz==3.10.1
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.26.1
sentence-transformers==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spotify==0.10.2
spotipy==2.24.0
sqlparse==0.5.2
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.40.0
sweetviz==2.3.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
visions==0.7.6
wcwidth==0.2.13
websocket-client==1.8.0
wordcloud==1.9.3
wsproto==1.2.0
xgboost==2.1.0
yarl==1.18.0
ydata-profiling==4.8.3
